CytokineProfileShinyApp: Comphrensive Cytokine Profiling Analysis Shiny Application

An interactive Shiny application that provides comprehensive cytokine profiling analysis through quality control using biologically meaningful cutoffs on raw cytokine measurements and by testing for distributional symmetry to recommend appropriate transformations. Offers exploratory data analysis with summary statistics, enhanced boxplots, and barplots, along with univariate and multivariate analytical capabilities for in-depth cytokine profiling such as Principal Component Analysis based on Andrzej Maćkiewicz and Waldemar Ratajczak (1993) doi:10.1016/0098-3004(93)90090-R , Sparse Partial Least Squares Discriminant Analysis based on Lê Cao K-A, Boitard S, and Besse P (2011) doi:10.1186/1471-2105-12-253 , Random Forest based on Breiman, L. (2001) doi:10.1023/A:1010933404324 , and Extreme Gradient Boosting based on Tianqi Chen and Carlos Guestrin (2016) doi:10.1145/2939672.2939785 .

See also

Author

Maintainer: Shubh Saraswat shubh.saraswat00@gmail.com (ORCID) [copyright holder]

Authors:

Xiaohua Douglas Zhang douglas.zhang@uky.edu (ORCID)